refactor: moved helper function to geomhelp
