+++ processing/processing.go
@@ -19,20 +19,6 @@ func readFeaturesFromSource(source Source, features chan<- Feature) {
 	source.ReadFeatures(features)
 }
 
-func PolygonSliceToGeom(polygons []geom.Polygon) geom.Geometry {
-	if len(polygons) == 0 {
-		panic("Multipolygon with zero polygons encountered")
-	}
-	if len(polygons) == 1 {
-		return polygons[0]
-	}
-	multipolygon := make(geom.MultiPolygon, len(polygons))
-	for i, p := range polygons {
-		multipolygon[i] = p
-	}
-	return multipolygon
-}
-
 func PolygonsToMulti(polygons []geom.Polygon) geom.MultiPolygon {
 	l := len(polygons)
 	multiPolygon := make(geom.MultiPolygon, l)
